+++ graph/graph-application/src/test/kotlin/eu/tib/orkg/prototype/statements/domain/model/SearchStringTest.kt
@@ -10,29 +10,39 @@ internal class SearchStringTest {
         @Nested
         inner class Query {
             @Test
-            fun `Escapes special characters`() {
-                FuzzySearchString("""+-&|!(){}[]^"~*?:\""").query shouldBe """\+\-\&\|\!\(\)\{\}\[\]\^\"\~\*\?\:\\"""
+            fun `Removes special characters`() {
+                FuzzySearchString("""+-&|!(){}[]^"~*?:\""").query shouldBe ""
             }
 
             @Test
             fun `Allows + and - when used as a word prefix`() {
-                FuzzySearchString("""+abc.""").query shouldBe """+abc."""
-                FuzzySearchString("""-def.""").query shouldBe """-def."""
+                FuzzySearchString("""+abc""").query shouldBe """+abc"""
+                FuzzySearchString("""-def""").query shouldBe """-def"""
             }
 
             @Test
             fun `Inserts AND operator between words`() {
-                FuzzySearchString("""some. label.""").query shouldBe """some. AND label."""
+                FuzzySearchString("""some label""").query shouldBe """*some* AND *label*"""
             }
 
             @Test
-            fun `Inserts wildcards after words`() {
-                FuzzySearchString("""some label""").query shouldBe """some* AND label*"""
+            fun `Inserts wildcards before and after words`() {
+                FuzzySearchString("""some label""").query shouldBe """*some* AND *label*"""
             }
 
             @Test
-            fun `Removes dashes when used within a word`() {
-                FuzzySearchString("""few-shot""").query shouldBe """few* AND shot*"""
+            fun `converts the given string correctly`() {
+                FuzzySearchString("""* ?irus few/zero-shot - abc +def: -ghi""").query shouldBe "*irus* AND *few* AND *zero* AND *shot* AND *abc* AND +def AND -ghi"
+            }
+
+            @Test
+            fun `Parses single character strings correctly`() {
+                FuzzySearchString("""a""").query shouldBe "*a*"
+            }
+
+            @Test
+            fun `Parses blank blank strings correctly`() {
+                FuzzySearchString("""    """).query shouldBe ""
             }
         }