+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added an alternative endpoint to query statements with a given predicate and literal called `/api/statements/predicate/{predicate_id}/literals/?q={literal}`
   (See: [!578](https://gitlab.com/TIBHannover/orkg/orkg-backend/-/merge_requests/578))
+### Changed
+- Fuzzy search by label now returns more results
+  (See: [!579](https://gitlab.com/TIBHannover/orkg/orkg-backend/-/merge_requests/579))
 
 ## [0.34.1] - 2023-06-06
 ### Fixed